change default temporary path
In case of installation on Windows, having absolute path "/tmp" as default value causes that CSS, JS files are not generated properly and the theme is not rendered correctly. 
I propose to set the default value to a relative path so that it exists also in case of installation on windows.